migrate percentile to use refactor
